MAINT: Use fixture in WithRecordAlgorithm.
