+++ zipline/test_algorithms.py
@@ -241,18 +241,6 @@ class TimeoutAlgorithm(TradingAlgorithm):
         pass
 
 
-class RecordAlgorithm(TradingAlgorithm):
-    def initialize(self):
-        self.incr = 0
-
-    def handle_data(self, data):
-        self.incr += 1
-        self.record(incr=self.incr)
-        name = 'name'
-        self.record(name, self.incr)
-        record(name, self.incr, 'name2', 2, name3=self.incr)
-
-
 class TestOrderAlgorithm(TradingAlgorithm):
     def initialize(self):
         self.incr = 0
